test the objects we parse, not reconstructions
The old way to test email parsing was:

1. parse input
2. make a new address object from pieces expected as parse output
3. compare parsed input to newly created address

In other words, we never checked that $parse->phrase eq $expected,
but instead did this:

  my ($parse) = Email::Address->parse($input);
  my $expect  = Email::Address->new(@expecting);
  is($parse->phrase, $expect->phrase);
  ...

This was a problem, because new sometimes massages input.  For
example, if you try to give the phrase q{"Bob"} your object's
phrase is q{Bob}.  This is good, but we want our tests to
reflect reality, not just pseudo-round-tripping.