unescape quoted phrase parts
The phrase of: "Bob \"Robert\" Smith" <bob@example.com>

should be q{Bob "Robert" Smith"} and not, as we got,
q{Bob \"Robert\" Smith}, which caused things not to round trip
properly.

It is possible that this changes how bogus input is parsed, and we
should add tests for that.  It improves parsing of valid input,
though.