Bring regexes more in line with RFC5322.
FWS is still brought in as \s+ for simplicity. The obs_phrase test in
comment passes fine, so remove that code; space backtracking problems
will be fixed in the next commit.

This commit actually fixes the original example in #10 as $comment is
much simplified.

Update test to pass; test is actually incorrect in that the "comment"
should not be removed (RT#80665).